refactor: extract IssueCoreSubtypes.swift from IssueCore.swift
Extracts 4 support types (ExtractedIssueCategory enum, ExtractedIssueSeverity
enum, IssueReproductionStep struct, IssueScreenshotAnnotation struct) from
IssueCore.swift into a new IssueCoreSubtypes.swift.

Byte-identical move; SHA-256 7271289bf235257970c70dec89de89801c05670722092b30599ed9cd4fc74c09
verified against origin/main:IssueCore.swift lines 3-147.

Reduces IssueCore.swift from 353 to 210 lines. Only the primary ExtractedIssue
struct remains in IssueCore.swift.

Tests: 667 (unchanged).
